refactor: merge scraping manager into scraping service
