+++ backend/src/main/java/it/compare/backend/scraping/scheduler/ScrapingScheduler.java
@@ -1,6 +1,6 @@
 package it.compare.backend.scraping.scheduler;
 
-import it.compare.backend.scraping.scraper.ScrapingManager;
+import it.compare.backend.scraping.service.ScrapingService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
@@ -9,10 +9,10 @@ import org.springframework.stereotype.Component;
 @RequiredArgsConstructor
 public class ScrapingScheduler {
 
-    private final ScrapingManager scrapingManager;
+    private final ScrapingService scrapingService;
 
     @Scheduled(cron = "0 0 */8 * * *")
-    public void startScraping() {
-        scrapingManager.scrapeAll();
+    public void run() {
+        scrapingService.scrapeAll();
     }
 }
